Added a couple of changes
