+++ PhotoEnhancer/PhotoEnhancer.csproj
@@ -53,6 +53,7 @@
     <Compile Include="Filters\IFilter.cs" />
     <Compile Include="Filters\LighteningFilter.cs" />
     <Compile Include="Filters\ParameterInfo.cs" />
+    <Compile Include="Filters\PixelFilter.cs" />
     <Compile Include="MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
